br-general-python/requirements/base.txt
# FastAPI and dependencies
fastapi==0.116.2
pydantic==2.11

# Pydantic settings management
pydantic-settings>=2.5

# ASGI server (to run FastAPI)
uvicorn[standard]==0.30.6

# Async HTTP client
httpx[http2]==0.28.1
requests==2.32.3

# Prisma ORM
prisma==0.15.0

# Environment variables
python-dotenv==1.0.1

# Email sending
aiosmtplib==4.0.2
Jinja2==3.1.6
email-validator==2.3.0

# User Security
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
passlib[argon2]==1.7.4

